pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "crassify"
version = "0.1.0"
description = "Protein-based viral taxonomy tool"
requires-python = ">=3.8"
dependencies = [
    "snakemake",
    "pandas",
    "tqdm",
    "rich",
    "pyyaml",
]

[tool.setuptools.packages.find]
include = ["crassify"]

[project.scripts]
crassify = "crassify.cli:main"

[tool.setuptools.package-data]
"crassify" = ["data/*"]
